Fleshed out + made copy of main script
Added a main script, fleshed it out, and finished adding portion that requests for Events, Categories, and Sources. Also added logic for HTTP requests in case some fail.

Was originally going to alter logging so that absolute paths are not shown for files in exception tracebacks. After some research, there are only kludge methods, disabling logging at certain portions of code, or altering a traceback's stack. It's complex to modify an existing exception's traceback or to create a new traceback from an existing StackSummary object, the other solutions are inadequate, so for now I'm leaving this as a potential issue to resolve later.

Made a copy of the code; I will be modifying pullData.py to see if I can skip merging Events, Categories, and Sources into a single .geojson file and instead to download each separately and connect them on Tableau. If the latter is possible then I'll remove pullDataOrig.py to make the script simpler.